feat: Added 'What should I do button?' to choose an option for the user
There has been a button added to the App.js that is responsible for
choosing an option out of the user's created list. This button also
has a disabled feature that allows the button to be clicked if there
are options but disabled if the user has not added any options to
the list.